refactor: move gtm token to tracking store
