+++ src/app/extensions/tracking/exports/tracking-exports.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { Angulartics2Module } from 'angulartics2';
+
+import { LAZY_FEATURE_MODULE } from 'ish-core/utils/module-loader/module-loader.service';
+
+@NgModule({
+  imports: [Angulartics2Module.forRoot()],
+  providers: [
+    {
+      provide: LAZY_FEATURE_MODULE,
+      useValue: { feature: 'tracking', location: import('../store/tracking-store.module') },
+      multi: true,
+    },
+  ],
+})
+export class TrackingExportsModule {}
