+++ src/app/extensions/tracking/store/tracking-config/tracking-config.actions.ts
@@ -1,0 +1,8 @@
+import { createAction } from '@ngrx/store';
+
+import { payload } from 'ish-core/utils/ngrx-creators';
+
+export const setGTMToken = createAction(
+  '[Configuration] Set Google Tag Manager Token',
+  payload<{ gtmToken: string }>()
+);
