+++ src/app/extensions/tracking/store/tracking-config/tracking-config.reducer.ts
@@ -1,0 +1,8 @@
+import { createReducer, on } from '@ngrx/store';
+
+import { setGTMToken } from './tracking-config.actions';
+
+export const trackingConfigReducer = createReducer<string>(
+  undefined,
+  on(setGTMToken, (_, action) => action.payload.gtmToken)
+);
