+++ src/app/extensions/tracking/store/tracking-config/tracking-config.selectors.ts
@@ -1,0 +1,5 @@
+import { createSelector } from '@ngrx/store';
+
+import { getTrackingState } from '../tracking-store';
+
+export const getGTMToken = createSelector(getTrackingState, state => state?.gtmToken);
