+++ src/app/extensions/tracking/store/tracking-store.ts
@@ -1,0 +1,7 @@
+import { createFeatureSelector } from '@ngrx/store';
+
+export interface TrackingState {
+  gtmToken: string;
+}
+
+export const getTrackingState = createFeatureSelector<TrackingState>('tracking');
